refactor: move `src/test/generate` directory to `src/generate`
