+++ src/generate/cpp/headers.ts
