+++ src/generate/cpp/utility.test.ts
